2 changes are happening in this commit:
1) libmysqlclient has been replaced for the supported mysqli, enabling
the upgrade of php's supported version from 4, to 5 and 7. This has only
been tested on version 5 for now.

2) Dockerfiles to build the application containerized. In order for that
to happen, there needs to be an image with php5 _and_ mysqli . The files
under pack/docker/php are aimed at building that, along with providing
required php.ini configuration within the docker container

In order to build this dependency container with mysqli, cd into
pack/docker/php an run:

$ docker build -t php5-mysqli .

This will generate an image called php5-mysqli , which is the base image
the root's Dockerfile is based on.

Afterwards, come down to the root of the git repo and build the
application image with (this step will require your docker cluster to
contain already the image php5-mysqli):

$ docker build -t inmersa-ierp .

to run the container with the application, just run:

$ docker run -d -p 80:80 -v /home/tite/proyectos/inmersa/Inmersa-IERP/vhost:/var/www/html --name ierp-dev inmersa-ierp

where the volume mapping on the container's endpoint /var/www/html
points to the vhost folder of the root for the repository.

Tite